+++ src/tests/news.test.ts
@@ -1,0 +1,25 @@
+import { connection } from "mongoose";
+import News from "@api/notice";
+
+const newObj = new News("test");
+const metadata = {
+  inLanguage: "en-us",
+  accessMode: "textual",
+  thumbnailUrl: null,
+  keywords: ["keyword"],
+};
+
+it("should write the new", () =>
+  expect(
+    newObj.write("Title", "# Content", "Description", "", metadata)
+  ).resolves.not.toBeNull());
+
+it("should return the new", () => expect(newObj.get()).resolves.not.toBeNull());
+
+it("should remove the new", () =>
+  expect(newObj.remove()).resolves.toBeUndefined());
+
+afterAll((done) => {
+  connection.close();
+  done();
+});