+++ src/utils/hash.js
@@ -1,8 +1,0 @@
-function hashString(str){
-    return require('crypto')
-        .createHash('sha256')
-        .update(str)
-        .digest('hex');
-}
-
-module.exports = { hashString };